requirements.txt
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
joblib
python-dotenv